finished stat upgrade for stats page in book ui

+++ ISP3_TEAM9/Assets/Scripts/Player/ScriptablePlayerStats.cs
@@ -36,12 +36,12 @@ public class playerStatPoints
 
 public class playerStatMultipliers
 {
-    public int health;
-    public int defense;
-    public int attackPower;
-    public int attackSpeed;
-    public int moveSpeed;
-    public int projectileSpeed; //For mage and archer
+    public float health;
+    public float defense;
+    public float attackPower;
+    public float attackSpeed;
+    public float moveSpeed;
+    public float projectileSpeed; //For mage and archer
 }
 
 [CreateAssetMenu]
